[FIX] hr_expense: add total amount label in specific case in expense form view
Updated the label for `total_amount` to display `Total` when the expense
is not in multiple currencies and the product has a cost, and to show an
empty label in the other previously considered cases. This change
ensures that the label is displayed correctly without repetitions or
omissions in the relevant scenarios.

closes odoo/odoo#183280

Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>